requirements.txt
streamlit>=1.26.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=1.5.0
plotly>=5.13.0
pillow>=9.4.0
matplotlib>=3.7.0
sympy>=1.11.0 